+++ src/main/java/com/example/tripKo/domain/place/api/AddressController.java
@@ -1,4 +1,4 @@
-package com.example.tripKo.domain.address.api;
+package com.example.tripKo.domain.place.api;
 
 import org.springframework.web.bind.annotation.RestController;
