+++ src/main/java/com/example/tripKo/domain/place/application/PlaceService.java
@@ -1,10 +1,10 @@
 package com.example.tripKo.domain.place.application;
 
 import com.example.tripKo._core.errors.exception.Exception404;
-import com.example.tripKo.domain.place.dto.response.PlaceFestivalResponse;
-import com.example.tripKo.domain.place.dto.response.PlaceResponse;
-import com.example.tripKo.domain.place.dto.response.PlaceRestaurantResponse;
-import com.example.tripKo.domain.place.dto.response.PlaceTouristSpotResponse;
+import com.example.tripKo.domain.place.dto.response.search.PlaceFestivalResponse;
+import com.example.tripKo.domain.place.dto.response.search.PlaceResponse;
+import com.example.tripKo.domain.place.dto.response.search.PlaceRestaurantResponse;
+import com.example.tripKo.domain.place.dto.response.search.PlaceTouristSpotResponse;
 import com.example.tripKo.domain.place.entity.PlaceFestival;
 import com.example.tripKo.domain.place.dao.PlaceFestivalRepository;
 import com.example.tripKo.domain.place.entity.PlaceRestaurant;
