+++ src/main/java/com/example/tripKo/domain/place/dao/AddressCategoryRepository.java
@@ -1,6 +1,6 @@
-package com.example.tripKo.domain.address.dao;
+package com.example.tripKo.domain.place.dao;
 
-import com.example.tripKo.domain.address.entity.AddressCategory;
+import com.example.tripKo.domain.place.entity.AddressCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AddressCategoryRepository extends JpaRepository<AddressCategory, Long> {
