+++ src/main/java/com/example/tripKo/domain/place/dao/AddressRepository.java
@@ -1,6 +1,6 @@
-package com.example.tripKo.domain.address.dao;
+package com.example.tripKo.domain.place.dao;
 
-import com.example.tripKo.domain.address.entity.Address;
+import com.example.tripKo.domain.place.entity.Address;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AddressRepository extends JpaRepository<Address, Long> {
