+++ src/main/java/com/example/tripKo/domain/place/dao/ContentsFileRepository.java
@@ -1,6 +1,6 @@
-package com.example.tripKo.domain.file.dao;
+package com.example.tripKo.domain.place.dao;
 
-import com.example.tripKo.domain.file.entity.ContentsHasFile;
+import com.example.tripKo.domain.place.entity.ContentsHasFile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ContentsFileRepository extends JpaRepository<ContentsHasFile, Long> {
