+++ src/main/java/com/example/tripKo/domain/place/dao/ContentsMenuRepository.java
@@ -1,6 +1,6 @@
-package com.example.tripKo.domain.contents.dao;
+package com.example.tripKo.domain.place.dao;
 
-import com.example.tripKo.domain.contents.entity.ContentsMenu;
+import com.example.tripKo.domain.place.entity.ContentsMenu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ContentsMenuRepository extends JpaRepository<ContentsMenu, Long> {
