+++ src/main/java/com/example/tripKo/domain/place/dao/ContentsRepository.java
@@ -1,6 +1,6 @@
-package com.example.tripKo.domain.contents.dao;
+package com.example.tripKo.domain.place.dao;
 
-import com.example.tripKo.domain.contents.entity.Contents;
+import com.example.tripKo.domain.place.entity.Contents;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ContentsRepository extends JpaRepository<Contents, Long> {
