+++ src/main/java/com/example/tripKo/domain/place/entity/AddressCategory.java
@@ -1,4 +1,4 @@
-package com.example.tripKo.domain.address.entity;
+package com.example.tripKo.domain.place.entity;
 
 import com.example.tripKo.domain.BaseTimeEntity;
 import lombok.*;
