+++ src/main/java/com/example/tripKo/domain/place/entity/Contents.java
@@ -1,4 +1,4 @@
-package com.example.tripKo.domain.contents.entity;
+package com.example.tripKo.domain.place.entity;
 import static javax.persistence.FetchType.LAZY;
 import static javax.persistence.GenerationType.IDENTITY;
 import static lombok.AccessLevel.PROTECTED;
@@ -15,8 +15,6 @@ import javax.persistence.OneToMany;
 import javax.persistence.Table;
 
 import com.example.tripKo.domain.BaseTimeEntity;
-import com.example.tripKo.domain.file.entity.ContentsHasFile;
-import com.example.tripKo.domain.place.entity.Place;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
